LayoutTests/imported/w3c/web-platform-tests/uievents/order-of-events/mouse-events/mouseover-out-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented
Steps:

Move your mouse over the blue <div> element, later over the green one, later over the yellow one.
Move the mouse from the yellow element to the green one, later to the blue one, and later over this paragraph.

Harness Error (FAIL), message = Unhandled rejection: unimplemented

NOTRUN Mouseover/out events
